edit hapus

+++ app/Http/Controllers/CabangController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
-
+use App\Models\Barang;
 use Illuminate\Http\Request;
 use App\Models\Cabang;
+use App\Models\Stok;
 
 class CabangController extends Controller
 {
@@ -31,17 +32,33 @@ class CabangController extends Controller
 
     public function delete($id)
     {
-        Cabang::find($id)->delete();
-        return redirect()->route('cabang.index')->with('success', 'Data Berhasil Dihapus');
+        $asset = Stok::where('cabang_id', $id)->first();
+        if (empty($asset)) {
+            Cabang::find($id)->delete();
+            return redirect()->route('cabang.index')->with('success', 'Data Berhasil Dihapus');
+        } else {
+            return redirect()->route('cabang.index')->with('error', 'Data Gagal dihapus karena sudah ada data transaksinya');
+        }
     }
 
-    public function laporan()
+    public function getEdit(Request $r)
     {
         $data = [
-            'title' => 'Laporan Data Cabang',
-            'cabang' => Cabang::orderBY('id', 'desc')->get()
-        ];
+            'title' => 'Edit User',
+            'cabang' => Cabang::find($r->id),
 
-        return view('laporan.cabang.index', $data);
+        ];
+        return view('superadmin.cabang.edit', $data);
+    }
+
+    public function update(Request $r)
+    {
+        $data = [
+            'nama' => $r->nama,
+            'alamat' => $r->alamat,
+            'ket' => $r->ket
+        ];
+        Cabang::where('id', $r->id)->update($data);
+        return redirect()->route('cabang.index')->with('success', 'Data Berhasil Disimpan');
     }
 }
